+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Not released
+
+- Add: Relative path could be used for `Backup Path`
+
 ## v0.5.3 (2021-04-03)
 
 - Add: Backup settings.json